requirements.txt
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
google-auth==2.40.2
greenlet==3.2.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
requests==2.32.3
rsa==4.9.1
SQLAlchemy==2.0.40
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
